Add files via upload
The motive of this project is to build a platform of football penalty shots. When the score depends on the number of goals within 5 attempts. The user controls the shot by giving horizontal and vertical positions. Besides, the goalkeeper acts as a bot to catch the ball by generating randomized positions and movement. After checking the match, the code generates the goal count and shows the score in the scoreboard. When the game finishes the program gets closed.